+++ tests/testthat/test-demean.R
@@ -64,7 +64,7 @@ test_that("demean shows message if some vars don't exist", {
 })
 
 
-test_that("demean for cross-classified designs (by > 1)", {
+test_that("demean for cross-classified designs (by > 1)", { #520
   skip_if(getRversion() < "4.1.0") # for pipe
 
   data(efc, package = "datawizard")