+++ NEWS.md
@@ -5,6 +5,11 @@
 * `data_select()` can directly rename selected variables when a named vector
   is provided in `select`, e.g. `data_select(mtcars, c(new1 = "mpg", new2 = "cyl"))`.
 
+* `data_tabulate()` gains an `as.data.frame()` method, to return the frequency
+  table as a data frame. The structure of the returned object is a nested data
+  frame, where the first column contains name of the variable for which
+  frequencies were calculated, and the second column contains the frequency table.
+
 # datawizard 0.11.0
 
 BREAKING CHANGES