+++ NEWS.md
@@ -1,4 +1,4 @@
-# datawizard 0.11.0.1
+# datawizard 0.11.1
 
 BREAKING CHANGES
 
@@ -20,6 +20,10 @@ CHANGES
   table as a data frame. The structure of the returned object is a nested data
   frame, where the first column contains name of the variable for which
   frequencies were calculated, and the second column contains the frequency table.
+
+* `demean()` (and `degroup()`) now also work for cross-classified designs, or
+  more generally, for data with multiple grouping or cluster variables (i.e.
+  `by` can now specify more than one variable).
 
 # datawizard 0.11.0